Add files via upload
The main objective of this project is to do a hypothesis test to check if the state has more vulnerable population (ages 65+) then there will be more influenza related deaths. We need this information to determine the number of staff to be sent to states with high death rates. The requirements were gathered, designed data research, sourced the right data from CDC and US census, data quality measures were taken to ensure there are no duplicates and the data is clean, complete, accurate and finally statistical analysis and hypothesis testing was done.